+++ e2e/fake-timers-legacy/reset-all-mocks/__tests__/resetAllMocks.test.js
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+test('works before resetAllMocks is called', () => {
+  jest.useFakeTimers();
+  const f = jest.fn();
+  setTimeout(f, 0);
+  jest.runAllTimers();
+  expect(f).toHaveBeenCalledTimes(1);
+});
+
+test('works after resetAllMocks is called', () => {
+  jest.resetAllMocks();
+  jest.useFakeTimers();
+  const f = jest.fn();
+  setTimeout(f, 0);
+  jest.runAllTimers();
+  expect(f).toHaveBeenCalledTimes(1);
+});
