+++ e2e/fake-timers-legacy/reset-mocks/__tests__/resetMock.test.js
@@ -5,4 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-module.exports = () => {};
+test('works when resetMocks is set in Jest config', () => {
+  jest.useFakeTimers();
+  const f = jest.fn();
+  setTimeout(f, 0);
+  jest.runAllTimers();
+  expect(f).toHaveBeenCalledTimes(1);
+});
