--- e2e/timer-reset-mocks/after-reset-all-mocks/package.json
@@ -1,6 +1,0 @@
-{
-  "jest": {
-    "testEnvironment": "node",
-    "resetMocks": false
-  }
-}
