--- e2e/timer-reset-mocks/with-reset-mocks/package.json
@@ -1,7 +1,0 @@
-{
-  "jest": {
-    "testEnvironment": "node",
-    "resetMocks": true,
-    "timers": "fake"
-  }
-}
